+++ src/index.js
@@ -22,6 +22,7 @@ const { errorResponse, base64 } = require('./utils.js');
 const github = require('./github');
 const google = require('./google');
 const onedrive = require('./onedrive');
+const { getCredentials } = require('./credentials.js');
 
 const DEFAULT_FORWARD_HEADERS = [
   'x-request-id',
@@ -99,11 +100,12 @@ async function main(req, context) {
           headers[header.toLocaleLowerCase()] = req.headers.get(header);
         }
         return headers;
-      }, {
-        // pass on authorization token
-        Authorization: gitHubToken ? `token ${gitHubToken}` : undefined,
-      }),
+      }, {}),
     };
+    if (gitHubToken) {
+      // pass on authorization token
+      githubOptions.headers.authorization = `token ${gitHubToken}`;
+    }
 
     const externalOptions = {
       GOOGLE_DOCS2MD_CLIENT_ID,
@@ -120,6 +122,7 @@ async function main(req, context) {
       || req.headers.get('x-openwhisk-activation-id')
       || '',
       namespace,
+      gitHubToken,
     };
 
     const dataOptions = {
@@ -187,6 +190,11 @@ async function main(req, context) {
       return errorResponse(log, 501,
         `No handler found for type ${mp.type} at path ${path} (${owner}/${repo})`,
         'Invalid mount configuration');
+    }
+
+    // extract credentials
+    if (gitHubToken && mp) {
+      externalOptions.credentials = getCredentials(log, mp, gitHubToken);
     }
 
     if (path.match(/sitemap(-[^/]+)?\.xml$/) && handler.handleSitemapXML) {