+++ src/onedrive.js
@@ -15,20 +15,15 @@ const { handleJSON } = require('./onedrive-json.js');
 const { handleSitemapXML } = require('./onedrive-sitemap.js');
 const { fetch, getFetchOptions, errorResponse } = require('./utils');
 const PenaltyBox = require('./penalty-box');
+const { getAccessToken } = require('./onedrive-helpers.js');
 
 // keep this global
 const box = new PenaltyBox();
 
 /**
  * Retrieves a file from OneDrive
- * @param {object} opts options
- * @param {object} opts.mp the mountpoint as defined by helix-shared
- * @param {string} opts.owner the GitHub org or username
- * @param {string} opts.repo the GitHub repository
- * @param {string} opts.ref the GitHub ref
- * @param {object} opts.log a Helix-Log instance
- * @param {object} opts.options Helix Fetch options
- * @param {Resolver} opts.resolver Version lock helper
+ * @param {ExternalHandlerOptions} opts the options
+ * @return {Promise<Response>} the http response
  */
 async function handle(opts) {
   const {
@@ -55,7 +50,13 @@ async function handle(opts) {
   url.searchParams.append('rid', options.requestId);
   url.searchParams.append('src', `${owner}/${repo}/${ref}`);
 
-  const response = await fetch(url.href, getFetchOptions(options));
+  const fetchOptions = getFetchOptions(options);
+
+  const accessToken = await getAccessToken(opts);
+  if (accessToken) {
+    fetchOptions.headers.authorization = `Bearer ${accessToken}`;
+  }
+  const response = await fetch(url.href, fetchOptions);
   const body = await response.text();
   if (response.ok) {
     const headers = {